#149 refactor for enabling auto updates. Moved update logic from API router to dedicated module
